# fixed/openapi.yaml
openapi: 3.0.0
info:
  title: Federation API
  version: 1.0.0
  description: |
    API de gestion d'une fédération agricole nationale.
    Inclut les fonctionnalités A, B-2, J, C/D, statistiques G/H.
  contact:
    name: Fédération Agricole
    email: contact@federation-agricole.mg
    url: https://federation-agricole.mg
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080/api/v1
    description: Serveur de développement
  - url: https://api.federation-agricole.mg/v1
    description: Serveur de production

tags:
  - name: Collectivités
  - name: Membres
  - name: Paiements
  - name: Comptes
  - name: Activités
  - name: Présences
  - name: Statistiques

paths:
  # ========== COLLECTIVITÉS ==========
  /collectivities:
    post:
      tags: [Collectivités]
      summary: Créer une collectivité (A)
      operationId: createCollectivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectivityRequest'
      responses:
        '201':
          description: Collectivité créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectivityResponse'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Collectivités]
      summary: Lister toutes les collectivités
      operationId: findAllCollectivities
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED]
        - name: city
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des collectivités
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectivityResponse'

  /collectivities/{id}:
    get:
      tags: [Collectivités]
      summary: Obtenir une collectivité par ID
      operationId: findCollectivityById
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
      responses:
        '200':
          description: Collectivité trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectivityResponse'
        '404':
          description: Non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /collectivities/{id}/validate:
    post:
      tags: [Collectivités]
      summary: Valider une collectivité par la fédération
      operationId: validateCollectivity
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
      responses:
        '200':
          description: Validée
        '404':
          description: Non trouvée

  /collectivities/{id}/informations:
    put:
      tags: [Collectivités]
      summary: Attribuer numéro et nom unique (J)
      operationId: assignCollectivityIdentity
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectivityInformationRequest'
      responses:
        '200':
          description: Informations mises à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectivityResponse'
        '400':
          description: Déjà attribué ou doublon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collectivité non trouvée

  /collectivities/{id}/membershipFees:
    get:
      tags: [Paiements]
      summary: Liste des cotisations d'une collectivité
      operationId: getMembershipFees
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
      responses:
        '200':
          description: Cotisations actives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembershipFeeResponse'
        '404':
          description: Collectivité non trouvée

    post:
      tags: [Paiements]
      summary: Créer des cotisations pour une collectivité
      operationId: createMembershipFees
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMembershipFeeRequest'
      responses:
        '201':
          description: Cotisations créées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembershipFeeResponse'
        '400':
          description: Données invalides
        '404':
          description: Collectivité non trouvée

  /collectivities/{id}/transactions:
    get:
      tags: [Paiements]
      summary: Transactions d'une collectivité sur une période
      operationId: getCollectivityTransactions
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Liste des transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectivityTransactionResponse'
        '400':
          description: Paramètres invalides
        '404':
          description: Collectivité non trouvée

  /collectivities/{id}/financialAccounts:
    get:
      tags: [Comptes]
      summary: Comptes financiers d'une collectivité avec solde à une date
      operationId: getFinancialAccounts
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
        - name: at
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date à laquelle obtenir le solde
      responses:
        '200':
          description: Liste des comptes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FinancialAccountResponse'
        '404':
          description: Collectivité non trouvée

  /collectivities/{id}/statistics:
    get:
      tags: [Statistiques]
      summary: Statistiques de paiement par membre (G)
      operationId: getCollectivityPaymentStats
      parameters:
        - $ref: '#/components/parameters/CollectivityId'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Statistiques de la collectivité
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectivityStatsResponse'

  /collectivities/statistics:
    get:
      tags: [Statistiques]
      summary: Statistiques globales de la fédération (H)
      operationId: getFederationStats
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Statistiques fédérales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationStatsResponse'

  # ========== MEMBRES ==========
  /members:
    post:
      tags: [Membres]
      summary: Créer un membre (B-2)
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          description: Membres créés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /members/{id}/join:
    post:
      tags: [Membres]
      summary: Valider l'adhésion d'un membre (B-2)
      operationId: joinMember
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Adhésion validée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          description: Conditions non remplies

  /members/{id}/transfer:
    post:
      tags: [Membres]
      summary: Transférer un membre
      operationId: transferMember
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [collectivityId]
              properties:
                collectivityId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Transfert réussi
        '400':
          description: Erreur

  /members/{id}/leave:
    post:
      tags: [Membres]
      summary: Faire quitter un membre
      operationId: leaveMember
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Départ enregistré

  /members/{id}/payments:
    post:
      tags: [Paiements]
      summary: Enregistrer des paiements d'un membre
      operationId: createMemberPayments
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMemberPaymentRequest'
      responses:
        '201':
          description: Paiements enregistrés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberPaymentResponse'
        '400':
          description: Données invalides

  # ========== COMPTES GÉNÉRAUX (déjà existants) ==========
  /accounts:
    post:
      tags: [Comptes]
      summary: Créer un compte
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Compte créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
    get:
      tags: [Comptes]
      summary: Lister tous les comptes
      operationId: findAllAccounts
      responses:
        '200':
          description: Liste des comptes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

  /accounts/{id}:
    get:
      tags: [Comptes]
      summary: Obtenir un compte par ID
      operationId: findAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Compte trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Non trouvé

  # ========== ACTIVITÉS & PRÉSENCES (bonus) ==========
  /activities:
    post:
      tags: [Activités]
      summary: Créer une activité
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '201':
          description: Activité créée
    get:
      tags: [Activités]
      summary: Lister les activités
      operationId: findAllActivities
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [MANDATORY, EXCEPTIONAL]
      responses:
        '200':
          description: Liste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityResponse'

  /attendances:
    post:
      tags: [Présences]
      summary: Enregistrer une présence
      operationId: createAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRequest'
      responses:
        '201':
          description: Présence enregistrée

  /activities/{activityId}/attendances:
    get:
      tags: [Présences]
      summary: Présences d'une activité
      operationId: findAttendancesByActivity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Liste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceResponse'

  # ========== STATISTIQUES MEMBRES (bonus) ==========
  /statistics/members/{id}:
    get:
      tags: [Statistiques]
      summary: Statistiques d'un membre
      operationId: getMemberStats
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Statistiques du membre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberStats'

components:
  parameters:
    CollectivityId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    # ========== Enums ==========
    CollectivityStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED]
    MemberStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]
    PaymentType:
      type: string
      enum: [MONTHLY, ANNUAL, EXCEPTIONAL]
    AccountType:
      type: string
      enum: [CASH, BANK, MOBILE_MONEY]
    ActivityType:
      type: string
      enum: [MANDATORY, EXCEPTIONAL]
    Frequency:
      type: string
      enum: [WEEKLY, MONTHLY, ANNUALLY, PUNCTUALLY]
    PaymentMode:
      type: string
      enum: [CASH, MOBILE_BANKING, BANK_TRANSFER]
    Gender:
      type: string
      enum: [MALE, FEMALE]

    # ========== Requêtes ==========
    CollectivityRequest:
      type: object
      required: [location, members, federationApproval, structure]
      properties:
        location:
          type: string
        specialty:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberIdentifier'
        federationApproval:
          type: boolean
        structure:
          $ref: '#/components/schemas/CreateCollectivityStructure'

    CreateCollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/MemberIdentifier'
        vicePresident:
          $ref: '#/components/schemas/MemberIdentifier'
        treasurer:
          $ref: '#/components/schemas/MemberIdentifier'
        secretary:
          $ref: '#/components/schemas/MemberIdentifier'

    CollectivityInformationRequest:
      type: object
      required: [number, name]
      properties:
        number:
          type: string
        name:
          type: string

    CreateMemberRequest:
      type: object
      required: [firstName, lastName, birthDate, phone, address, referees]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        profession:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        collectivityId:
          type: string
          format: uuid
        referees:
          type: array
          items:
            $ref: '#/components/schemas/SponsorRequest'

    SponsorRequest:
      type: object
      required: [sponsorId, relationType]
      properties:
        sponsorId:
          type: string
          format: uuid
        relationType:
          type: string
          enum: [FAMILY, FRIEND, COLLEAGUE, NEIGHBOR, OTHER]
        relationDetails:
          type: string

    CreateMembershipFeeRequest:
      type: object
      required: [eligibleFrom, frequency, amount, label]
      properties:
        eligibleFrom:
          type: string
          format: date
        frequency:
          $ref: '#/components/schemas/Frequency'
        amount:
          type: number
        label:
          type: string

    CreateMemberPaymentRequest:
      type: object
      required: [amount, accountCreditedIdentifier, paymentMode]
      properties:
        amount:
          type: number
        membershipFeeIdentifier:
          type: string
        accountCreditedIdentifier:
          type: string
        paymentMode:
          $ref: '#/components/schemas/PaymentMode'

    AccountRequest:
      type: object
      required: [name, type, collectivityId]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        collectivityId:
          type: string
          format: uuid

    ActivityRequest:
      type: object
      required: [title, activityDate, type, collectivityId]
      properties:
        title:
          type: string
        description:
          type: string
        activityDate:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/ActivityType'
        collectivityId:
          type: string
          format: uuid

    AttendanceRequest:
      type: object
      required: [activityId, memberId, present]
      properties:
        activityId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        present:
          type: boolean
        justification:
          type: string

    # ========== Réponses ==========
    CollectivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        name:
          type: string
        location:
          type: string
        specialty:
          type: string
        creationDate:
          type: string
          format: date
        validatedByFederation:
          type: boolean
        status:
          $ref: '#/components/schemas/CollectivityStatus'
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberResponse'

    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        joinDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/MemberStatus'
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string
        referees:
          type: array
          items:
            $ref: '#/components/schemas/SponsorResponse'

    SponsorResponse:
      type: object
      properties:
        sponsorId:
          type: string
          format: uuid
        sponsorName:
          type: string
        relationType:
          type: string
        relationDetails:
          type: string

    MembershipFeeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eligibleFrom:
          type: string
          format: date
        frequency:
          $ref: '#/components/schemas/Frequency'
        amount:
          type: number
        label:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]

    MemberPaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
        paymentMode:
          $ref: '#/components/schemas/PaymentMode'
        accountCreditedId:
          type: string
          format: uuid
        creationDate:
          type: string
          format: date

    FinancialAccountResponse:
      oneOf:
        - $ref: '#/components/schemas/CashAccountResponse'
        - $ref: '#/components/schemas/MobileBankingAccountResponse'
        - $ref: '#/components/schemas/BankAccountResponse'

    CashAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [CASH]
        amount:
          type: number

    MobileBankingAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [MOBILE_MONEY]
        holderName:
          type: string
        mobileBankingService:
          type: string
        mobileNumber:
          type: string
        amount:
          type: number

    BankAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [BANK]
        holderName:
          type: string
        bankName:
          type: string
        bankCode:
          type: string
        bankBranchCode:
          type: string
        bankAccountNumber:
          type: string
        bankAccountKey:
          type: string
        amount:
          type: number

    CollectivityTransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creationDate:
          type: string
          format: date
        amount:
          type: number
        paymentMode:
          $ref: '#/components/schemas/PaymentMode'
        accountCredited:
          $ref: '#/components/schemas/FinancialAccountResponse'
        memberDebited:
          $ref: '#/components/schemas/MemberResponse'

    CollectivityStatsResponse:
      type: object
      properties:
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string
        memberStats:
          type: array
          items:
            $ref: '#/components/schemas/MemberPaymentStats'

    MemberPaymentStats:
      type: object
      properties:
        memberId:
          type: string
          format: uuid
        memberName:
          type: string
        totalPaid:
          type: number
        unpaids:
          type: number

    FederationStatsResponse:
      type: object
      properties:
        period:
          type: string
        upToDateStats:
          type: array
          items:
            $ref: '#/components/schemas/CollectivityUpToDateStat'
        newMembersStats:
          type: array
          items:
            $ref: '#/components/schemas/CollectivityNewMembersStat'

    CollectivityUpToDateStat:
      type: object
      properties:
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string
        upToDatePercentage:
          type: number

    CollectivityNewMembersStat:
      type: object
      properties:
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string
        newMembersCount:
          type: integer

    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        isActive:
          type: boolean
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string

    ActivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        activityDate:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/ActivityType'
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string

    AttendanceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        memberName:
          type: string
        present:
          type: boolean
        justification:
          type: string

    CollectivityStats:
      type: object
      properties:
        collectivityId:
          type: string
          format: uuid
        collectivityName:
          type: string
        totalMembers:
          type: integer
        activeMembers:
          type: integer
        averageAttendanceRate:
          type: number
        totalPayments:
          type: number
        paymentsByMonth:
          type: object
          additionalProperties:
            type: number

    MemberStats:
      type: object
      properties:
        memberId:
          type: string
          format: uuid
        memberName:
          type: string
        seniorityDays:
          type: integer
        totalPayments:
          type: number
        activitiesAttended:
          type: integer
        attendanceRate:
          type: number
        mandateHistory:
          type: array
          items:
            $ref: '#/components/schemas/MandateInfo'

    MandateInfo:
      type: object
      properties:
        roleName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # Utilitaire
    MemberIdentifier:
      type: string